Finshed FindFloorCoordinate function using the tranform class
Former-commit-id: 1a6f37fd98dff518806dbcbc096ba38ec0f74dd8 [formerly be1b764b64ae39d88a9f32fca167f1e7f5cba607]
Former-commit-id: cf01278893b3a384b55c4c55706b5a40857064b1

+++ soccerbot/soccer_fieldline_detection/include/soccer_fieldline_detection/camera.hpp
@@ -39,11 +39,11 @@ public:
 
     float ImageSensorWidth();
 
-    int PixelHeight();
+    float PixelHeight();
 
-    int PixelWidth();
+    float PixelWidth();
 
-    int ImageSensorLocation_X(int pos_x, int pos_y);
+    float ImageSensorLocation_X(int pos_x, int pos_y);
 
-    int ImageSensorLocation_Y(int pos_x, int pos_y);
+    float ImageSensorLocation_Y(int pos_x, int pos_y);
 };
